Split up `use_self` ui test
